refactor(FavoritePlaceRow): Rename. Absorb NewFavoriteLocation.
