+++ lib/components/user/favorite-locations/favorite-locations-pane.js
@@ -3,8 +3,7 @@ import React from 'react'
 import { ControlLabel } from 'react-bootstrap'
 
 import { isHomeOrWork } from '../../../util/user'
-import FavoriteLocation from './favorite-location'
-import NewFavoriteLocation from './new-favorite-location'
+import FavoritePlaceRow from './favorite-place-row'
 
 /**
  * Renders an editable list user's favorite locations, and lets the user add a new one.
@@ -22,18 +21,19 @@ const FavoriteLocationsPane = ({ values: userData }) => {
           <>
             {savedLocations.map((loc, index) => {
               return (
-                <FavoriteLocation
+                <FavoritePlaceRow
                   arrayHelpers={arrayHelpers}
                   index={index}
                   isFixed={isHomeOrWork(loc)}
                   key={index}
                   location={loc}
+                  onClick={null}
                 />
               )
             })}
 
             {/* For adding a location. */}
-            <NewFavoriteLocation arrayHelpers={arrayHelpers} />
+            <FavoritePlaceRow onClick={null} />
           </>
         )}
       />
